Introduce the generic single object event handler
NewSingleObjectEventHandler creates a new event handler for a single object.
It exploits a WithWatch client to watch for changes on the object and execute a goroutine to listen for the channel
and call the handler function when an event occurs.
The function is generic and takes two types T and L, where T is the type of the object to watch and L is the type of the list of T objects to watch. The function also takes the name of the object the handler should subscribe to and the namespace to watch.
handler is a function that takes the event type and the object that was changed. Event types are defined in watch.go and can be Added, Modified, Deleted, Bookmark and Error (not handled by handler).
errorHandler is an optional (nullable pointer to a) function executed when the event type is Error.