Fix permissions to access images.config.openshift.io and the openshift-image-registry/image-registry-certificates ConfigMap
This commit adds the permissions to access images.config.openshift.io cluster-wide and the image-registry-certificates ConfigMap in the openshift-image-registry namespace.

Adding cluster-wide and namespaced permissions seems not a well-supported feature by the kube-builder. Therefore, the namespaced role and rolebinding are added manually in the kustomization folders.

Moreover, since the default namespace transformer is used to replace the namespace in all the resources, but the namespaced Role and RoleBinding have to land in a different namespace than the operator one, this commit also adds an overlay to allow resources in specific, hard-coded namespaces and avoid their replacement when the kustomization is applied.

Kustomize is also updated to 4.5.7