Update the Dockerfile to allow replacing base images via build-args and use public golang builder in registry.ci
This is needed more for envs different than Prow, and enable users to control the base images to use from the Dockerfile.

Also note that this is pinning to the use of a single-arch builder image currently as we don't have any public manifest list image for the builders. The current runtime stage multi-arch images built via docker-buildx are unaffected.