Do not cache the image inspection results until an eviction policy is implemented
The current caching mechanism comes from the PoC implementation but is not ready yet to be used because (1) it misses a cache eviction policy, (2) it may need a few further fixes to avoid leaking metadata of images when the inspection happens with different pull secrets.

This commit skip the use of the cacheProxy to solely rely on the remote registry inspector.