Improve logging by allowing changes to the level at runtime
We reconcile the operands when a change to the ClusterPodPlacementConfig is detected. That reconciliation also included the setup of the log-level with the default zap command line flag.
However, the operator didn't change the level of logging.
This commit changes this behavior by leveraging a custom command line flag and the zap.AtomicLevel object that can be changed at runtime.
When a CusterPodPlacementConfig is reconciled, the controllers continue to be reconciled as we've always done, but the operator can now change the log level at runtime.
Moreover, the dev mode is now disabled so that the logs output uses the default non-dev JSON encoder that allows collection of logs through structured log collectors.

# Conflicts:
#	main.go