MULTIARCH-4934: Ordered removal of operand
When the ClusterPodPlacementConfig is removed, the Pod placement controller should wait for the webhook to be deleted and should ungate all the pods.

This commit will let the operator add a finalizer to the ClusterPodPlacementConfig and the pod placement controller deployment such that when the user deletes the ClusterPodPlacementConfig, the deletion is marked for all the objects, with the webhook immediately collected to avoid new pods from being gated. The pod placement controller will continue running until all the pods are ungated. Then, the operator removes the finalizer in the deployment and the ClusterPodPlacementConfig, allowing the operand to be disabled without leaking any gated pod in the cluster.