Send events to the Pods to better inform about the actions of the scheduling gate
Not all the users can be allowed to look at the logs of the pod placement controller or the web hook. Having information about the pods being scheduling gated and the result of the images inspection through events associated to them (also visible in K8S UIs) will improve the UX for users of this operator.