MULTIARCH-4934: Re-queue the delete handling by firing an error if the webhook service is still available before checking for the pods gating status
We can consider the service existence as a gate to the end of communication between the API server and the mutating webhook handler.
If the server still exists when we check for any scheduling gate left in our pods, we might incur in a race condition where the webhook gates a pod just after we retrieve the list of pods in Pending phase. Therefore, we remove the finalizers and the controller may not be in time to remove the scheduling gate.

Therefore, before checking for the pods gating status, we try to retrieve the service. If it's not found, we move forward, otherwise we return an error to re-queue the deletion handling. The non-existence of the service ensures no communication between the API server and any, still-active, webhook pod.