MULTIARCH-4937: Implement Status of the ClusterPodPlacementConfig, conditions and improve reliability.
This commit implements the handling of the following conditions for the ClusterPodPlacementConfig:
   - Degraded: if some components are not available (no replicas) and the object is not deprovisioning
   - Deprovisioning: if the object is being deleted
   - MutatingWebhookConfigurationNotAvailable: if the mutating webhook configuration does not exist
   - PodPlacementControllerNotReady: if the pod placement controller is not available or up-to-date
   - PodPlacementWebhookNotReady: if the pod placement webhook is not available or up-to-date
   - Progressing: if the object is not deprovisioning and some of the components are not up-to-date.
   - Available: if all the components are available to serve the requests and reconcile node affinities (at least one replica).

This allows the users to be informed about the status of the operand.
When the operand is available and all the components are up-to-date, the only condition shown/set is "Available."

When the ClusterPodPlacementConfig is updated, or some components are being rolled out, the "Progressing" condition is set.

When progressing, if at least one replica is available for both the controller and webhook, the operand is considered available.
In this case, we also allow the deployment of the mutating webhook configuration.

Despite the current FailurePolicy of the MutatingWebhookConfiguration is set to None, if some components lack at least one replica's availability, we forcibly delete the MutatingWebhookConfiguration, and the Degraded condition is set. Similarly, we now deploy the MutatingWebhookConfiguration only if at least one replica of the MutatingWebhook and Controller deployments are available (such that we can ensure both the admission patching and the removal of gates through the controller).

When removing the object, the deprovisioning condition is set and the users are informed about the controller pending deletion if there exists at least one pod that is still gated.