OCPBUGS-38601: The operand should not fail when the operator is uninstalled without prior deletion of the operator
When the operator is uninstalled without prior deletion of the operand, we should not let the operand to fail. This is due to the removal of the service account and other RBAC objects. This commit will let the operator to create the RBAC objects and additional service accounts, detached from the OLM ones.
This commmit also includes an edge-case fix where a deployment is deleted outside of the lifecycle of the operand: the operator must be able to reconcile it, therefore we need to remove the finalizer.