Do not schedule pods when the images supported architectures sets are disjoint
The pod placement controller faces an error when updating the pod because the matchSelector we compute is not valid. It cannot have an empty list of values (kubernetes.io/arch IN [] is not valid). As we just remove the scheduling gate in case of errors, the pods entered the scheduling cycle without modifications.

When the images of a pod do not have a commonly supported architecture, we now add a dummy node selector requirement that will let the pod to hold in the Unschedulable status.