OCPBUGS-38823: skip bundle images when computing the set of compatible architectures for a given pod
This commit updates the registry inspector to handle bundle images correctly. Bundle images are intended to carry metadata-only information and not be multi-architecture.

It is possible for an operator bundle image to reference a valid operator supported by the cluster's architectures while the bundle image itself remains a single-architecture manifest image. Suppose the cluster (e.g., an arm64+ppc64le cluster) has no nodes supporting the bundle image's architecture (such as an amd64-built bundle). OLM may fail to run the job in that case because the node affinity settings could lead to an ImagePullBackOff error.

To address this, the GetCompatibleArchitectureSet function will now return a set with all the supported architectures (Arm64, Amd64, Ppc64le, S390x) for OLM bundle images. If there are other container images specified in the pod (as in the OLM case), the node affinity will be set to the intersection of the architectures supported by the non-bundle images. As this fix is strictly related to the OLM bundles, we don't consider the case where a pod has only that bundle as a container image. In fact, the pod cannot run in such a case because no binaries are shipped in the image of a bundle to act as an entry point.