The pod reconciler should also run the shouldIgnore verification
This change ensures that we remove the scheduling gate for pods that no longer meet the criteria for architecture-aware scheduling due to changes in-between the webhook patch at creation time and the processing of the pod by the pod-placement-controller.

- A pod may be gated but unprocessed when the operator's configuration changes, making the pod ineligible for architecture-aware scheduling based on the new rules.
- Alternatively, the pod may have been modified by another webhook in the admission chain (e.g., setting the `nodeAffinity` for the `kubernetes.io/arch` label), rendering it unsuitable for further processing by the operator.

In both scenarios, we must not set the nodeAffinity and the gate should be removed.