Ignore pods that have nodeSelectors/affinity set earlier
This commit adds a method that check if the pod has already a nodeSelector for the architecture label or if all the nodeSelectorTerms in the nodeAffinity field have a matchExpression for the architecture label.
The method is run by the pod placement controller and by the webhook, so that we can prevent the webhook from patching such pods at all.