The cache of the ClusterPodPlacementConfig controller should filter by status.phase only for the pods resource
The pod reconciler should reconcile only pods that are in status.phase==pending. However, other informers and reconcilers in the same controller that watch for other resources (for example, the CPPCInformer) should not get the same filter.
This commit changes the cache options to have a byObject filter instead of a default filter for all the watched resources.