Scaffolding for the enoexec-daemon
This commit implements the initial scaffolding for the enoexec-daemon as an additional gomodule. The CI scripts are updated to run for any additional modules and the Containerfiles now include a build of the enoexec-daemon in the runtime image.

In the future, we may move this binary to a dedicated container image.

Signed-off-by: aleskandro <aleskandro@redhat.com>