Implement Kubernetes storage for ENOEXEC events
This commit introduces K8sENOExecEventStorage as a storage backend that creates ENoExecEvent custom resources in the Kubernetes cluster.

The backend runs in its own goroutine and consumes data sent through a channel of ENOEXECInternalEvent objects. The expected writer for this channel is the eBPF tracepoint component.

When an event is received, this storage implementation creates a corresponding ENoExecEvent object in the cluster. To prevent API server and key-value store flooding, especially in scenarios such as CrashLoopBackOff or incorrect binaries retrying in a loop, a token bucket-based rate limiter (golang.org/x/time/rate) is used to throttle requests.

Additionally, each event processing operation is wrapped in a timeout context. This defends against prolonged blocking or denial-of-service (DoS) conditions under high load or malicious input.

Throttling parameters are currently hardcoded and will be refined or made configurable during system fine-tuning.

Note: this is a first implementation of a storage backend and that, as discussed in the EP, we might implement a backend storage that leverages an IPC mechanism to seaparate more the event generation and the creation of the K8S CRs by deploying the two components as a two-containers DaemonSet.

Signed-off-by: aleskandro <aleskandro@redhat.com>