[Refactoring] Reorganize repo struct to use a single module (vendor)
Initially, we have configured the structure of the enoexec daemon as a separate go module. However, this creates circular dependencies and looks overkilling for the case and not beneficial. This commit reorganizes the structure of the repo, moving the main.go files in the cmd/ folder and integrating the enoexec event daemon code in the already available packages.

We will continue to build two separate binaries, but this solution is simpler and easier to maintain.